Added code coverage
