+++ settings.json
@@ -1,5 +1,8 @@
 {
     "debug": false,
     "hubUri": "wss://microservicebus.com",
-    "port": 80
+    "nodeName": "TestNode1",
+    "organizationId": "65b22e1f-a17e-432f-b9f2-b7057423a786",
+    "port": 9090,
+    "machineName": "CND5215QH6"
 }
